Consolidate Orders list UI ownership
