+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/orders/list/OrderListTestTags.kt
@@ -1,6 +1,25 @@
 package com.woocommerce.android.ui.orders.list
 
 internal object OrderListTestTags {
+    const val SCREEN = "order_list_screen"
+    const val BODY = "order_list_body"
+    const val SEARCH_ACTION = "order_list_search_action"
+    const val BARCODE_ACTION = "order_list_barcode_action"
+    const val SEARCH_FIELD = "order_list_search_field"
+    const val BROWSING_CONTROLS = "order_list_browsing_controls"
+    const val FILTERS = "order_list_filters"
+    const val JITM = "order_list_jitm"
+    const val CREATE_ORDER_FAB = "order_list_create_order_fab"
+    const val TROUBLESHOOTING = "order_list_troubleshooting"
+    const val TROUBLESHOOTING_TOGGLE = "order_list_troubleshooting_toggle"
+    const val TROUBLESHOOTING_ACTION = "order_list_troubleshooting_action"
+    const val CONTACT_SUPPORT_ACTION = "order_list_contact_support_action"
+    const val SELECTION_HEADER = "order_list_selection_header"
+    const val SELECTION_CLOSE = "order_list_selection_close"
+    const val SELECTION_TITLE = "order_list_selection_title"
+    const val SELECTION_OVERFLOW = "order_list_selection_overflow"
+    const val SELECTION_MENU = "order_list_selection_menu"
+    const val SELECTION_UPDATE_STATUS = "order_list_selection_update_status"
     const val LIST = "order_list"
     const val INITIAL_LOADING = "order_list_initial_loading"
     const val EMPTY = "order_list_empty"
