+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/orders/list/OrderListUiModels.kt
@@ -2,6 +2,49 @@ package com.woocommerce.android.ui.orders.list
 
 import androidx.annotation.ColorRes
 
+internal data class OrderListScreenState(
+    val isSearchActive: Boolean = false,
+    val searchQuery: String = "",
+    val filterCount: Int = 0,
+    val lastUpdate: String? = null,
+    val rowState: OrderListRowState = OrderListRowState(),
+    val troubleshooting: OrderListTroubleshootingPresentation? = null,
+    val showCreateOrderFab: Boolean = true,
+) {
+    val isSelecting: Boolean
+        get() = rowState.isBulkSelectionActive
+
+    val selectedOrderCount: Int
+        get() = rowState.bulkSelectedOrderIds.size
+
+    private val headerMode: OrderListHeaderMode
+        get() = when {
+            isSelecting -> OrderListHeaderMode.Selection
+            isSearchActive -> OrderListHeaderMode.Search
+            else -> OrderListHeaderMode.Browsing
+        }
+
+    val headerContent: OrderListHeaderContent
+        get() = OrderListHeaderContent(
+            mode = headerMode,
+            selectedOrderCount = selectedOrderCount,
+        )
+
+    val shouldShowCreateOrderFab: Boolean
+        get() = showCreateOrderFab && headerMode != OrderListHeaderMode.Selection
+}
+
+internal data class OrderListHeaderContent(
+    val mode: OrderListHeaderMode,
+    val selectedOrderCount: Int,
+)
+
+internal enum class OrderListHeaderMode {
+    Selection,
+    Search,
+    Browsing,
+}
+
 internal sealed interface OrderListContentState {
     data object InitialLoading : OrderListContentState
 
@@ -57,4 +100,14 @@ internal data class OrderListRowState(
 ) {
     val isBulkSelectionActive: Boolean
         get() = bulkSelectedOrderIds.isNotEmpty()
+}
+
+internal data class OrderListTroubleshootingPresentation(
+    val type: OrderListTroubleshootingType,
+    val isExpanded: Boolean = true,
+)
+
+internal enum class OrderListTroubleshootingType {
+    ParsingError,
+    Timeout,
 }
